cancel appEngine saga on RESET_EDITOR_REQUEST
